refactor: bootstrap plugin via main class
